Populate and use TestResult. option to show error.
use existing model that didnt get populated with test results. only initialized to default values. 
Show error message underneith aswell, but added option so we can turn it off. 
but usually we want to know what the error is and not only the name(just like when we run the tests and print out the error, now we see them at the bottom summary too)