fix(http,ws): conn typings on internal utils
Update the return type from `Deno.Conn` to `Deno.Conn<Deno.NetAddr>` on
two of the internal functions used to mock conn objects. The issue was
introduced in core at `9c7c9a35`.

In the case of `http/_mock_conn.ts@mockConn` I also had to add an
explicit type hint on the return object because of the `base` parameter
that was causing the type checker to not like the `transport` property
on `localAddr` and `remoteAddr`.