fix(hash): Partially mitigate WASM memory leak #786 using FinalizationRegistry
This change uses `FinalizationRegistry` (which is working correctly on deno@main as of today) to .free() the Hasher instances in WASM when their corresponding `Hash` JavaScript objects are garbage collected. This does **not** entirely resolve the memory leak: the WASM heap will never shrink and release the used memory, but it can significantly mitigate it. In a test on my machine, endlessly calling `createHash('sha256')`, the WASM heap grew more slowly, and stopped growing around ~70MB, as the allocator inside the WASM was able to reuse the heap space we freed. Without this change, the memory use continued to grow without bound; I stopped it after 1GB.

This won't be effective in all cases: The garbage collector may not get a chance to run if you're repeatedly creating hashes in a hot sync loop. If you're passing huge chunks of data into WASM, the heap will still need to be expanded to hold them. But I think it's still a significant improvement.

It appears that wasm-bindgen does have an option that would allow it to do this automatically internally, but it hasn't been exposed through wasm-pack that we're using here, and it's not clear that it will be soon.