fix(datetime): dayOfYear for the southern hemisphere
Because of the direction of time to/from daylight savings is
different for the northern and southern hemisphere the effect
of counting seconds between 1st of Jan and another date goes from
adding a few hours (not a problem when counting days) to
subtracting a few hours (problematic when you count days from midnight
and end up on the day before).

The solution proposed here is to disregard any timezone information
as the amount of days in UTC is the same as the amount of days in
any local timezone.