test(fs/util): separate URL cases by OS
In order to properly test the function on posix sysetms without
implementing a way to manually specify the separator in the function,
URLs using the win32 path style need to be filtered out.

There should never be a case where you would use win32 style paths on
a posix system, and those paths would resolve to the same location.