fix(http/file_server): resolve empty subdir correctly on Windows
One goal of this commit is to avoid tripping over the 'current directory read permission' prompt
from deno.
The prompt should not happen in the case when we already specified a path to the `--allow-read`
argument.
This prompt can be seen when navigating to a subdir in a browser while running a deno server
from Windows. This is because file_server mixed different path separators when building
paths and then pass it to `Deno.stat()`. However `Deno.stat()` doesn't seem to handle it well
and therefore generated the prompt.