refactor: remove `user.id` use and rework vote logic (#401)
Vote logic was challenging to deal with, hence the rework. The most
involved part of this PR is the migration script. In summary:
* All notifications will be deleted, which, IMO, is ok.
* User entries are untouched.
* Items and votes entries are migrated according to their new respective
interfaces.

Reviews, please pay close attention to the migration script. It aims to
preserve item, vote and user data. This script has worked multiple times
on my local machine, but I still want it checked carefully. To test:
1. On the main branch, run `deno task db:reset` and then `deno task
db:seed`.
2. Run `deno task start`, sign in, vote and comment on random items,
noting the changes.
3. Switch to this branch and run `deno run -A --unstable
tools/migrate_kv.ts`.
4. Run `deno task start` and ensure all data is preserved.

Please review, if willing and able, @brunocorrea23 and @mbhrznr.

Closes #400