feat: welcome page (#622)
The client is directed to the welcome page if GitHub (OAuth) needs to be
set up. This page directs the dev to the setup guides for local and
production setups. This change also allows GitHub environment variables
not to be defined upon startup. This aims to give a dev trying out
SaaSKit for the first time a straightforward guide on how to get
started.