fix(fs/ensure_dir): allow links to directories
`lstat()` will not consider a symlink to a directory as a directory.
Any use case of `ensureDir()` will likely consider a symlink to a directory
as a directory.

If a symlink to a directory needs to be replaced/handled differently,
it should be handled by the caller outside of `ensureDir()`.