Make debounce work with asynchronous function
## Motivation

The async package provides helper functions for asynchronous operations, yet the debounce function does not work with asynchronous functions. Since these functions are commonly the most valuable functions to debounce, It makes sense to design debounce to support asynchrony.

This also allows us to return a value from debounce.

## Solution

* Simplified the types for debuts: it's now able to infer the arguments and return type of the wrapped function.
* Every function call saves a resolver in an array: Every call to a debounced function will resolve the same result at the same time.

Caveat: I wrote this in the GitHub editor before bed, it probably has a type error I forgot about somewhere -- will fix it tomorrow, but worth a review in the meantime.